+++ index.html
@@ -233,7 +233,7 @@
   </template>
 
   <script scr="/js/functions.js"> </script>
-  <script scr="/js/main.js"> </script>
+  <script scr="/js/main.js" type="module"> </script>
 
 </body>
 </html>
